requirements.txt
# core
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0
faker==24.1.0

# ml
scikit-learn==1.4.1.post1
joblib==1.3.2
torch==2.2.1
shap==0.45.0

# viz
matplotlib==3.8.3

# api
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
anthropic==0.54.0

# tracking
mlflow==2.11.1

# env
python-dotenv==1.0.1
